+++ app/controllers/account_controller.rb
@@ -85,7 +85,8 @@ class AccountController < ApplicationController
 
   def set_user_type
     user_types = params.required(:user).permit(:provider, :researcher, :adult_diagnosed, :adult_at_risk, :caregiver_adult, :caregiver_child)
-    current_user.update user_types
+    current_user.update_user_types user_types
+
     redirect_to get_started_privacy_path
   end