+++ test/models/answer_session_test.rb
@@ -3,7 +3,7 @@ require "test_helper"
 class AnswerSessionTest < ActiveSupport::TestCase
 
   test "#grouped_reportable_answers" do
-    assert_not_empty answer_sessions(:complete).grouped_reportable_answers
+    assert_not_empty answer_sessions(:dep_complete).grouped_reportable_answers
   end