+++ test/models/survey_test.rb
@@ -12,7 +12,7 @@ class SurveyTest < ActiveSupport::TestCase
     assert_not_nil survey
 
     assert_equal "about-me", survey.slug
-    assert_equal 5, survey.all_questions_descendants.length
+    assert_equal 5, survey.all_questions_descendants.length, "hmm: #{survey.all_questions_descendants.map(&:slug)}"
 
     assert_match /What is your date of birth?/, survey.all_questions_descendants.first.text
 
@@ -38,17 +38,24 @@ class SurveyTest < ActiveSupport::TestCase
 
   end
 
+  test "All defined surveys should load with no problem" do
+    assert_difference "Survey.count", Survey::SURVEY_LIST.length do
+      Survey::SURVEY_LIST.each {|survey_slug| Survey.load_from_file(survey_slug)}
+    end
+
+  end
+
   test "self.migrate_old_answers" do
     question_map = [{"slug" => "favorite-color", "answer_template_name" => "color_list", "id" => questions(:q1).id.to_s }]
-    survey = Survey.find_by_slug("new-survey")
+    survey = Survey.find_by_slug("hidden-survey")
 
     assert_difference "AnswerSession.count", 2 do
-      Survey.migrate_old_answers("new-survey", question_map)
+      Survey.migrate_old_answers("hidden-survey", question_map)
     end
 
 
-    unmapped_question = questions(:new_q1)
-    mapped_question = questions(:new_q2)
+    unmapped_question = questions(:number2)
+    mapped_question = questions(:radio2)
 
     assert_equal 0, unmapped_question.answers.count
     assert_equal 2, mapped_question.answers.count
@@ -72,7 +79,7 @@ class SurveyTest < ActiveSupport::TestCase
     end
 
 
-    assert_equal u.email, surveys(:new).launch_single(u, "baseline")
+    assert_equal u, surveys(:new).launch_single(u, "baseline")
 
 
 
@@ -82,10 +89,10 @@ class SurveyTest < ActiveSupport::TestCase
 
     assert_difference "AnswerSession.where(encounter: 'baseline').count", User.current.where(adult_diagnosed: true).count do
       result = surveys(:new_2).launch_multiple(User.current.where("adult_diagnosed = TRUE"), 'baseline')
-      assert_match /previously launched: 0/, result
+      assert_empty result
     end
 
-    assert_match /previously launched: 18/, surveys(:new_2).launch_multiple(User.current.where("adult_diagnosed = TRUE"), 'baseline')
+    assert_equal User.current.where(adult_diagnosed: true), surveys(:new_2).launch_multiple(User.current.where("adult_diagnosed = TRUE"), 'baseline')
   end
 
 end